+++ scripts/Muon/GUI/Common/corrections_tab_widget/corrections_tab_widget.py
@@ -18,7 +18,8 @@ class CorrectionsTabWidget(object):
     def __init__(self, context, parent):
         self.corrections_tab_view = CorrectionsView(parent)
         self.corrections_tab_model = CorrectionsModel(context.data_context, context.corrections_context)
-        self.corrections_tab_presenter = CorrectionsPresenter(self.corrections_tab_view, self.corrections_tab_model)
+        self.corrections_tab_presenter = CorrectionsPresenter(self.corrections_tab_view, self.corrections_tab_model,
+                                                              context)
 
         context.update_view_from_model_notifier.add_subscriber(
             self.corrections_tab_presenter.update_view_from_model_observer)
